requirements.txt
streamlit
crewai
openai
pypdf2
python-dotenv
pandas
langchain-openai